Override default locale
The issue was that the `C.UTF-8` locale was being used, but `C.UTF-8`
doesn't exist on my system. We switch to a version of `rules_haskell`
that allows setting the locale, switch the locale to something that's on
my system, and it's fixed!

This isn't a very useful fix though. It means you have to know what's on
each system in order to know what locale to use. Ideally, `hGetContents`
wouldn't be privvy to locale issues and we wouldn't have to deal with
this at all. But, them the breaks.